moved web tree view files to common folder
