+++ packages/plugin-core/src/common/ITreeViewConfig.ts
