+++ packages/plugin-core/src/common/TreeNote.ts
