+++ packages/plugin-core/src/common/TreeViewDummyConfig.ts
