+++ packages/plugin-core/src/web/injection-providers/setupWebExtContainer.ts
@@ -17,8 +17,8 @@ import { NoteLookupProvider } from "../commands/lookup/NoteLookupProvider";
 import { DendronEngineV3Web } from "../engine/DendronEngineV3Web";
 import { VSCodeFileStore } from "../engine/store/VSCodeFileStore";
 import { PreviewPanel } from "../views/preview/PreviewPanel";
-import { ITreeViewConfig } from "../views/treeView/ITreeViewConfig";
-import { TreeViewDummyConfig } from "../views/treeView/TreeViewDummyConfig";
+import { ITreeViewConfig } from "../../common/ITreeViewConfig";
+import { TreeViewDummyConfig } from "../../common/TreeViewDummyConfig";
 import { getAssetsPrefix } from "./getAssetsPrefix";
 import { getEnablePrettlyLinks } from "./getEnablePrettlyLinks";
 import { getSiteIndex } from "./getSiteIndex";
