+++ packages/plugin-core/src/web/test/suite/injection-providers/setupWebExtContainer.test.ts
@@ -7,7 +7,7 @@ import { URI, Utils } from "vscode-uri";
 import { CopyNoteURLCmd } from "../../../commands/CopyNoteURLCmd";
 import { NoteLookupCmd } from "../../../commands/NoteLookupCmd";
 import { setupWebExtContainer } from "../../../injection-providers/setupWebExtContainer";
-import { NativeTreeView } from "../../../views/treeView/NativeTreeView";
+import { NativeTreeView } from "../../../../common/NativeTreeView";
 import { WorkspaceHelpers } from "../../helpers/WorkspaceHelpers";
 
 async function setupEnvironment() {
